+++ cpp/timing.h
@@ -20,7 +20,7 @@
 extern "C" {
 #endif
 double timefunc_s(void (*func)(uint32_t n, uint32_t k, uint32_t* result), uint32_t n, uint32_t k,
-                  uint32_t iters);
+                  uint64_t iters);
 #if __cplusplus
 }
 #endif
